Split cluster initialisation and client access.
For MPI clusters we want to be able to initialise the cluster early as all workers will
run the main code until dask_mpi is used to convert them into workers. Delaying
initialisation means they will all parse the config. I also think this is a cleaner
interface anyway.

Note that the context manager interface is removed with this commit. I think that
managing the lifecycle of the cluster should be left to the plugin and/or the code
managing the simulation, not the simulation method.