Add spec_source entry to plugin_spec dictionary.
This is optional, and should be a string or path giving the source of the plugin
specification. This may be used for error reporting, and plugins may use it e.g., to
resolve a filename to some data relative to the configuration file that specified the
plugin. If a plugin initialiser accepts a spec_source parameter, the plugin loading
functions in `openstb.simulator.plugin.loader` will automatically pass the value from
the spec (if it contains it).