plugin.util: load_config_plugins() function.
This function uses type annotations, either introspected from the desired configuration
specification or explicitly given. It goes through a configuration dictionary and loads
(with the appropriate registered loader function based on the type annotations) each
plugin within it. It also has error-checking options for missing or extra entries if
desired.