Point simulation: improve efficiency of cluster usage.
Instead of submitting tasks for one ping+receiver and waiting until they complete,
monitor the number of unfinished simulation tasks. When they drop below a threshold
proportional to the number of workers, submit the next set of tasks. This ensures there
are always tasks available and so should minimise the idle time of the workers.