Point sim: don't include targets beyond maximum range.
* Set amplitude of echoes from these targets to zero.

* Add a guard band to the end of the traces so the tails of echoes from targets near the
  maximum don't wrap around to the start of the trace.

This is probably subomptimal; we should be able to discard those targets earlier and not
apply all the effects to them. The receiver properties will be made more flexible (e.g.,
variable sample length to model systems with a variable ping rate) and the maximum range
handling will be improved as part of that. This may go into the travel time calculator
which could return a mask of which targets to include.