Iterative TWTT: clear error message if past end of trajectory
If a ping is transmitted close to the end of the trajectory, then the echoes from a
target may arrive after the end of the trajectory. In this case we cannot find the
position at the receiption point to proceed with our iterations.

In the previous code, this ended up failing to converge, raising the corresponding
exception. This commit changes it to a clearer exception so the user knows what the real
problem was.